requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
cryptography==41.0.4
qrcode[pil]==7.4.2
opencv-python==4.8.1.78
pyzbar==0.1.9
PyMuPDF==1.23.5
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0